move APIFnSend to action_send.go
